requirements.txt
fastapi
pydantic
pydantic-settings
python-dotenv
python-multipart
uvicorn
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
pypdf
sqlalchemy
asyncpg
psycopg2-binary
pyJWT
bcrypt
email-validator
langchain-huggingface 
sentence_transformers

